Servers on home screen
idea here is to move away from multiple servers and displaying all media mushed in together, this sets up a server button and allows us to pick a server to use.

once servers is selected(emby/plex) all content from that server will be added next to TvShow and Movie button for sqllite/mysql database(if avaialable) and inProgres and Recently media will only be displayed for that server in addition to local DB(if exists).

skins that dont have button 4000 (server button) on home screen are incompatible with this change and will fallback to old ways